[ParkingLot][Story6][Refactor] datatype change

+++ src/main/java/com/parkinglot/ParkingLot.java
@@ -53,8 +53,8 @@ public class ParkingLot {
         return vacancy;
     }
 
-    public int getAvailablePositionRate() {
-        return vacancy / capacity;
+    public double getAvailablePositionRate() {
+        return (double) vacancy / capacity;
     }
